requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
langchain-pinecone
pypdf # Para PDFs
boto3 # Para AWS S3
langchain-text-splitters # Asegura que RecursiveCharacterTextSplitter esté disponible
langchain-community # Para PyPDFLoader, TextLoader, UnstructuredMarkdownLoader
langchain-core # Para Document